Data.Colour.RGBSpace.HSV

Documentation

data RGB a

An RGB triple for an unspecified colour space.

hsvView :: (Fractional a, Ord a) => RGB a -> (a, a, a)

Returns the HSV (hue-saturation-value) coordinates of an RGB triple. See hue, saturation, and value.

hue :: (Fractional a, Ord a) => RGB a -> a

The hue coordinate of an RGB value is in degrees. Its value is always in the range 0-360.

saturation :: (Fractional a, Ord a) => RGB a -> a

Returns the saturation coordinate (range [0,1]) of an RGB triple for the HSV (hue-saturation-value) system. Note: This is different from saturation for the Data.Colour.RGBSpace.HSL

value :: (Fractional a, Ord a) => RGB a -> a

Returns the value coordinate (raonge [0,1]) of an RGB triple for the HSV (hue-saturation-value) system.

hsv :: (RealFrac a, Ord a) => a -> a -> a -> RGB a

Convert HSV (hue-saturation-value) coordinates to an RGB value. Hue is expected to be measured in degrees [0,360], while saturation and value are expected to be in the closed range [0,1].