Data.Colour.CIE

Description

Colour operations defined by the International Commission on Illumination (CIE).

Documentation

data Colour a

This type represents the human preception of colour. The a parameter is a numeric type used internally for the representation.

The Monoid instance allows one to add colours, but beware that adding colours can take you out of gamut. Consider using blend whenever possible.

cieXYZ :: Fractional a => a -> a -> a -> Colour a

Construct a Colour from XYZ coordinates for the 2° standard (colourimetric) observer.

cieXYZView :: Fractional a => Colour a -> (a, a, a)

Returns the XYZ colour coordinates for the 2° standard (colourimetric) observer.

luminance :: Fractional a => Colour a -> a

Returns the Y colour coordinate (luminance) for the 2° standard (colourimetric) observer.

toCIEXYZ :: Fractional a => Colour a -> (a, a, a)

Deprecated: toCIEXYZ has been renamed cieXYZView

data Chromaticity a

mkChromaticity :: Fractional a => a -> a -> Chromaticity a

Constructs Chromaticity from the CIE little x, little y coordinates for the 2° standard (colourimetric) observer.

chromaCoords :: Fractional a => Chromaticity a -> (a, a, a)

Returns the CIE little x, little y, little z coordinates for the 2° standard (colourimetric) observer.

chromaX :: Fractional a => Chromaticity a -> a

Returns the CIE little x coordinate for the 2° standard (colourimetric) observer.

chromaY :: Fractional a => Chromaticity a -> a

Returns the CIE little y coordinate for the 2° standard (colourimetric) observer.

chromaZ :: Fractional a => Chromaticity a -> a

Returns the CIE little z coordinate for the 2° standard (colourimetric) observer.

chromaConvert :: (Fractional b, Real a) => Chromaticity a -> Chromaticity b

Change the type used to represent the chromaticity coordinates.

chromaColour

Arguments

:: Fractional a

=> Chromaticity a

-> a
luminance

-> Colour a

Constructs a colour from the given Chromaticity and luminance.

lightness

Arguments

:: (Ord a, Floating a)

=> Chromaticity a
White point

-> Colour a

-> a

Returns the lightness of a colour with respect to a given white point. Lightness is a perceptually uniform measure.

cieLABView

Arguments

:: (Ord a, Floating a)

=> Chromaticity a
White point

-> Colour a

-> (a, a, a)

Returns the CIELAB coordinates of a colour, which is a perceptually uniform colour space. The first coordinate is lightness. If you don't know what white point to use, use d65.

cieLAB

Arguments

:: (Ord a, Floating a)

=> Chromaticity a
White point

-> a
L* coordinate (lightness)

-> a
a* coordinate

-> a
b* coordinate

-> Colour a

Returns the colour for given CIELAB coordinates, which is a perceptually uniform colour space. If you don't know what white point to use, use d65.

Orphan instances

AffineSpace Chromaticity

Methods

affineCombo :: Num a => [(a, Chromaticity a)] -> Chromaticity a -> Chromaticity a