Data.Colour.SRGB.Linear

Description

Provides a linear colour space with the same gamut as Data.Colour.SRGB.

Documentation

data Colour a

This type represents the human preception of colour. The a parameter is a numeric type used internally for the representation.

The Monoid instance allows one to add colours, but beware that adding colours can take you out of gamut. Consider using blend whenever possible.

data RGB a

An RGB triple for an unspecified colour space.

Constructors

RGB

Fields

channelRed :: !a

channelGreen :: !a

channelBlue :: !a

rgb :: Fractional a => a -> a -> a -> Colour a

Constructs a Colour from RGB values using the linear RGB colour with the same gamut as sRGB.

toRGB :: Fractional a => Colour a -> RGB a

Return RGB values using the linear RGB colour with the same gamut as sRGB.

sRGBGamut :: RGBGamut

This is the gamut for the sRGB colour space.